+++ ethereum/executionengine/src/main/java/tech/pegasys/teku/ethereum/executionengine/auth/JwtAuthWebsocketHelper.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2022 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.ethereum.executionengine.auth;
+
+import com.google.common.net.HttpHeaders;
+import java.util.Optional;
+import org.web3j.protocol.websocket.WebSocketClient;
+import tech.pegasys.teku.infrastructure.time.TimeProvider;
+
+public class JwtAuthWebsocketHelper {
+  private final SafeTokenProvider tokenProvider;
+  private final TimeProvider timeProvider;
+
+  public JwtAuthWebsocketHelper(final JwtConfig jwtConfig, final TimeProvider timeProvider) {
+    this.tokenProvider = new SafeTokenProvider(new TokenProvider(jwtConfig));
+    this.timeProvider = timeProvider;
+  }
+
+  public void setAuth(final WebSocketClient webSocketClient) {
+    final Optional<Token> optionalToken = tokenProvider.token(timeProvider.getTimeInMillis());
+    if (optionalToken.isEmpty()) {
+      return;
+    }
+    final Token token = optionalToken.get();
+    final String authHeader = "Bearer " + token.getJwtToken();
+    webSocketClient.addHeader(HttpHeaders.AUTHORIZATION, authHeader);
+  }
+}
