--- infrastructure/el/build.gradle
@@ -1,18 +1,0 @@
-dependencies {
-  implementation project(':infrastructure:time')
-  implementation project(':infrastructure:exceptions')
-  implementation project(':infrastructure:unsigned')
-  implementation project(':infrastructure:time')
-  implementation project(':infrastructure:logging')
-  implementation project(':infrastructure:async')
-
-  api 'org.web3j:core'
-  implementation 'org.apache.tuweni:tuweni-units'
-  implementation 'io.jsonwebtoken:jjwt-api'
-
-  runtimeOnly 'io.jsonwebtoken:jjwt-impl'
-  runtimeOnly 'io.jsonwebtoken:jjwt-jackson'
-
-  testImplementation testFixtures(project(':infrastructure:async'))
-  testImplementation testFixtures(project(':infrastructure:time'))
-}
