--- infrastructure/el/src/test/java/tech/pegasys/teku/infrastructure/el/Web3jClientBuilderTest.java
@@ -1,84 +1,0 @@
-/*
- * Copyright 2022 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.infrastructure.el;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
-
-import org.junit.jupiter.api.Test;
-import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
-import tech.pegasys.teku.infrastructure.time.TimeProvider;
-
-public class Web3jClientBuilderTest {
-  @Test
-  public void shouldFailSettingBadEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    assertThatThrownBy(() -> builder.endpoint("http://^"))
-        .isInstanceOf(InvalidConfigurationException.class);
-  }
-
-  @Test
-  public void shouldFailBuildPreconditionCheckWithNoEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    assertThatThrownBy(builder::build).isInstanceOf(NullPointerException.class);
-  }
-
-  @Test
-  public void shouldFailBuildWithUnsupportedEndpointScheme() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    builder.timeProvider(mock(TimeProvider.class)).endpoint("abc://localhost");
-    assertThatThrownBy(builder::build).isInstanceOf(InvalidConfigurationException.class);
-  }
-
-  @Test
-  public void shouldBuildHttpClientWithHttpEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    builder.timeProvider(mock(TimeProvider.class)).endpoint("http://localhost");
-    Web3JClient client = builder.build();
-    assertThat(client).isInstanceOf(Web3jHttpClient.class);
-  }
-
-  @Test
-  public void shouldBuildHttpClientWithHttpsEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    builder.timeProvider(mock(TimeProvider.class)).endpoint("https://localhost");
-    Web3JClient client = builder.build();
-    assertThat(client).isInstanceOf(Web3jHttpClient.class);
-  }
-
-  @Test
-  public void shouldBuildWebsocketClientWithWsEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    builder.timeProvider(mock(TimeProvider.class)).endpoint("ws://localhost");
-    Web3JClient client = builder.build();
-    assertThat(client).isInstanceOf(Web3jWebsocketClient.class);
-  }
-
-  @Test
-  public void shouldBuildWebsocketClientWithWssEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    builder.timeProvider(mock(TimeProvider.class)).endpoint("wss://localhost");
-    Web3JClient client = builder.build();
-    assertThat(client).isInstanceOf(Web3jWebsocketClient.class);
-  }
-
-  @Test
-  public void shouldBuildIpcClientWithFileEndpoint() {
-    Web3jClientBuilder builder = new Web3jClientBuilder();
-    builder.timeProvider(mock(TimeProvider.class)).endpoint("file:/some/path");
-    Web3JClient client = builder.build();
-    assertThat(client).isInstanceOf(Web3jIpcClient.class);
-  }
-}
