+++ services/executionengine/build.gradle
@@ -3,10 +3,10 @@ dependencies {
   implementation project(':infrastructure:events')
   implementation project(':infrastructure:exceptions')
   implementation project(':ethereum:executionlayer')
+  implementation project(':ethereum:executionengine')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:spec')
   implementation project(':infrastructure:async')
-  implementation project(':infrastructure:el')
   implementation project(':infrastructure:logging')
   implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:version')
