+++ services/powchain/build.gradle
@@ -3,9 +3,9 @@ dependencies {
   implementation project(':infrastructure:events')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:networks')
+  implementation project(':ethereum:executionengine')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:bls')
-  implementation project(':infrastructure:el')
   implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:logging')
   implementation project(':infrastructure:metrics')
