+++ settings.gradle
@@ -6,6 +6,7 @@ include 'ethereum:core'
 include 'ethereum:spec'
 include 'ethereum:dataproviders'
 include 'ethereum:events'
+include 'ethereum:executionengine'
 include 'ethereum:executionlayer'
 include 'ethereum:forkchoice'
 include 'ethereum:networks'
@@ -26,7 +27,6 @@ include 'infrastructure:bls'
 include 'infrastructure:bytes'
 include 'infrastructure:collections'
 include 'infrastructure:crypto'
-include 'infrastructure:el'
 include 'infrastructure:events'
 include 'infrastructure:exceptions'
 include 'infrastructure:http'
