+++ teku/build.gradle
@@ -6,13 +6,13 @@ dependencies {
   implementation project(':data:publisher')
   implementation project(':data:beaconrestapi')
   implementation project(':infrastructure:bls')
-  implementation project(':infrastructure:el')
   implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:metrics')
   implementation project(':data:dataexchange')
   implementation project(':data:yaml')
   implementation project(':ethereum:core')
   implementation project(':ethereum:dataproviders')
+  implementation project(':ethereum:executionengine')
   implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:pow:api')
